refactor: extract points redux slice from settings one to reduce the scope of data update effect
