+++ src/components/hooks/useStampCompleteBonus.ts
@@ -1,5 +1,6 @@
-import { settingsSelector, useStamps } from '../../store/settings'
 import { useSelector } from 'react-redux'
+import { settingsSelector } from '../../store/settings'
+import { useStamps } from '../../store/points'
 import { usePostApiV1ProfileByProfileIdPointMutation } from '../../generated/dreamkast-api.generated'
 import {
   getAllStampCollected,
